+++ iceoryx_examples/icedelivery/iox_publisher_typed_modern.cpp
@@ -14,7 +14,7 @@
 
 #include "topic_data.hpp"
 
-#include "iceoryx_posh/popo/modern_api/publisher.hpp"
+#include "iceoryx_posh/popo/publisher.hpp"
 #include "iceoryx_posh/runtime/posh_runtime.hpp"
 
 #include <chrono>
