+++ iceoryx_posh/include/iceoryx_posh/internal/popo/base_publisher.inl
