+++ iceoryx_posh/include/iceoryx_posh/internal/popo/typed_publisher.inl
