+++ iceoryx_posh/include/iceoryx_posh/popo/base_subscriber.hpp
@@ -16,7 +16,7 @@
 #define IOX_POSH_POPO_BASE_SUBSCRIBER_HPP
 
 #include "iceoryx_posh/internal/popo/ports/subscriber_port_user.hpp"
-#include "iceoryx_posh/popo/modern_api/sample.hpp"
+#include "iceoryx_posh/popo/sample.hpp"
 #include "iceoryx_posh/popo/wait_set.hpp"
 #include "iceoryx_posh/runtime/posh_runtime.hpp"
 #include "iceoryx_utils/cxx/expected.hpp"
@@ -174,6 +174,6 @@ class BaseSubscriber
 } // namespace popo
 } // namespace iox
 
-#include "iceoryx_posh/internal/popo/modern_api/base_subscriber.inl"
+#include "iceoryx_posh/internal/popo/base_subscriber.inl"
 
 #endif // IOX_POSH_POPO_BASE_SUBSCRIBER_HPP
