+++ iceoryx_posh/include/iceoryx_posh/popo/untyped_subscriber.hpp
@@ -17,7 +17,7 @@
 
 #include "iceoryx_posh/capro/service_description.hpp"
 #include "iceoryx_posh/iceoryx_posh_types.hpp"
-#include "iceoryx_posh/popo/modern_api/base_subscriber.hpp"
+#include "iceoryx_posh/popo/base_subscriber.hpp"
 #include "iceoryx_utils/cxx/expected.hpp"
 #include "iceoryx_utils/cxx/unique_ptr.hpp"
 
@@ -63,6 +63,6 @@ using UntypedSubscriber = UntypedSubscriberImpl<>;
 } // namespace popo
 } // namespace iox
 
-#include "iceoryx_posh/internal/popo/modern_api/untyped_subscriber.inl"
+#include "iceoryx_posh/internal/popo/untyped_subscriber.inl"
 
 #endif // IOX_POSH_POPO_UNTYPED_SUBSCRIBER_HPP
