+++ iceoryx_posh/test/moduletests/test_popo_base_publisher.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iceoryx_posh/popo/modern_api/base_publisher.hpp"
+#include "iceoryx_posh/popo/base_publisher.hpp"
 #include "mocks/chunk_mock.hpp"
 #include "mocks/publisher_mock.hpp"
