+++ iceoryx_posh/test/moduletests/test_popo_untyped_publisher.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iceoryx_posh/popo/modern_api/untyped_publisher.hpp"
+#include "iceoryx_posh/popo/untyped_publisher.hpp"
 #include "mocks/chunk_mock.hpp"
 #include "mocks/publisher_mock.hpp"
