+++ tools/introspection/include/iceoryx_introspection/introspection_app.hpp
@@ -15,7 +15,7 @@
 #define IOX_TOOLS_ICEORYX_INTROSPECTION_INTROSPECTION_APP_HPP
 
 #include "iceoryx_introspection/introspection_types.hpp"
-#include "iceoryx_posh/popo/modern_api/subscriber.hpp"
+#include "iceoryx_posh/popo/subscriber.hpp"
 #include "iceoryx_utils/platform/getopt.hpp"
 
 #include <map>
